+++ R/components.R
@@ -130,7 +130,7 @@ BaseComponent <- R6::R6Class(
     .forceCallArgs = NULL,
     .global = NULL,
     ## ---------------------------------------------------------------
-    trigger = function(name, data, force = TRUE) {
+    trigger = function(name, data = NULL, force = TRUE) {
       if (FALSE && force && self$inObserver()) {
         self$log("hack", name)
         private$.forceCall(
@@ -146,7 +146,15 @@ BaseComponent <- R6::R6Class(
       }
       if (name %in% ls(self$events)) {
         if (is.null(data)) {
-          self$events[[name]] <- shiny::isolate(!self$events[[name]])
+          self$events[[name]] <- shiny::isolate({
+            if (is.logical(self$events[[name]])) {
+              !self$events[[name]]
+            } else if (is.null(self$events[[name]]$value)) {
+              TRUE
+            } else {
+              !self$events[[name]]$value
+            }
+          })
         } else {
           data$timestamp <- as.numeric(Sys.time())*1000
           self$events[[name]] <- data