+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.4.0
 ### Features
 * eventEmitter based logging system
+* load function (for loading components in R files)
+### Bugfix
+* fix calling trigger without arg and with NULL after auto-creating event with `on`
 
 ## 0.3.1
 ### Bugfix
@@ -8,7 +11,7 @@
 
 ## 0.3.0
 ### Features
-* Event Emitters and services
+* Event Emitters and static services
 
 ## 0.2.0
 ### Features